further reduced flag time and flag warning time
36 hours seems like plenty of time to afford the opportunity for someone
to make a move.

there may be a bias in the system due to the fact that lost positions
are regularly passed/ignored, whereas games that end with checkmate are
taken out of the queue. I think I can alleviate this situation a bit by
reducing the flag time so that games nobody wants to play are cleaned up
faster.

see
https://www.reddit.com/r/chess/comments/c59g86/distrochess_official_release/es2k0gb?utm_source=share&utm_medium=web2x