bugfix: game.id == null
This is just a note to let you know that a game in which you
participated is about to be flagged. If you wish to continue playing for
your team, then you should make a move soon. Otherwise, this game will
be recorded as a loss for your team in 6 hours.
Click here to play this game (game ID: null).